requirements.txt
agents==1.4.0
openai==2.29.0
pydantic==2.12.5
python-dotenv==1.2.2
streamlit==1.45.1
